interface.py: for clean add a dryrun option (also helps coverage)

+++ interface.py
@@ -193,11 +193,12 @@ class Interface(Parser):
         else:
             print('status no bashrc notes in the settings')
 
-    def clean(self, sure=False):
+    def clean(self, sure=False, dryrun=False):
         """
         Remove all installed components from this folder.
         Use this command before reinstalling supporting software with
         the refresh command.
+        'dryrun' just helps us confirm the files
         """
         import shutil
         print('status cleaning')
@@ -210,11 +211,11 @@ class Interface(Parser):
         fns += [i for i in glob.glob('modulefiles/*') if i != 'modulefiles/cc']
         self.cache = {}
         print('status removing: %s' % ', '.join(fns))
-        if sure or confirm('okay to remove the files above?',):
-            for fn in fns:
-                shutil.rmtree(fn) if os.path.isdir(fn) else os.remove(fn)
-            # os.mkdir('tmp')
-            print('status done')
+        if not dryrun:
+            if sure or confirm('okay to remove the files above?',):
+                for fn in fns:
+                    shutil.rmtree(fn) if os.path.isdir(fn) else os.remove(fn)
+                print('status done')
 
     def showcache(self):
         """
